==> notebooks/03_feature_engineering.ipynb <==
import pandas as pd
import numpy as np
from datetime import timedelta


df = pd.read_csv('../data/processed/cleaned_transactions.csv', parse_dates=['InvoiceDate'])

df.head()


snapshot_date = df['InvoiceDate'].max() + timedelta(days=1)
snapshot_date


rfm = df.groupby('CustomerID').agg({
    'InvoiceDate': lambda x: (snapshot_date - x.max()).days,   # Recency
    'InvoiceNo': 'nunique',                                     # Frequency
    'TotalPrice': 'sum'                                         # Monetary
}).reset_index()

rfm.columns = ['CustomerID', 'Recency', 'Frequency', 'Monetary']
rfm.head()


additional_features = df.groupby('CustomerID').agg({
    'Quantity': ['sum', 'mean'],
    'TotalPrice': ['mean'],
    'InvoiceDate': ['min', 'max']
})

additional_features.columns = [
    'total_quantity',
    'avg_quantity',
    'avg_order_value',
    'first_purchase_date',
    'last_purchase_date'
]

additional_features = additional_features.reset_index()
additional_features.head()


customer_features = pd.merge(rfm, additional_features, on='CustomerID')
customer_features.head()


churn_threshold = 90  # days

customer_features['Churn'] = np.where(
    customer_features['Recency'] > churn_threshold,
    1,
    0
)

customer_features['Churn'].value_counts(normalize=True)


customer_features.info()


customer_features.head()


customer_features.to_csv('../data/processed/customer_features.csv', index=False)
